==> preprocessing/01_repos_metadata.ipynb <==
**Description:** In this notebook a file with the relevant metadata of multiple GitHub repositories will be created. Keep in mind, the files are already created! If you want to test the notebook create new filenames, for example *../data/df_repos_metadata_star_up_to_max_TEST.json*  
**Depencencies:**  
+ data/raw_data_zip/raw_data_no_range.zip  
+ data/raw_data_zip/raw_data_0_22196.zip  

*Hint: If lines are created with support of a Large Language Model or the code is taken from another source, you find following hint at the end of the line: (generated with Microsoft Copilot) or (source: link_to_source)*

import pandas as pd # package for data manipulation
import json # package to work with .json
from pathlib import Path # package for path manipulation
from langdetect import detect # package for language detection
import zipfile # package to work with ZIP files
import os # package to work with operating system

# custom function to load a json file for a given path
def load_json(path):
    with open(path, 'r') as file:
        loaded_data = json.load(file)
    
    return loaded_data

# custom function to rezip files for given path
def rezip_files(path, range_type):
    original_zip = path # path to the original ZIP file
    if range_type == 'range':
        extracted_dir = '../data/raw_data/range' # directory where the extract contents are saved
    else:
        extracted_dir = '../data/raw_data/no_range'

    with zipfile.ZipFile(original_zip, 'r') as zip_ref: 
        zip_ref.extractall(extracted_dir) # extract all files

# columns which are requiered for preprocessing and further steps
columns = [
    'id', 
    'name', 
    'full_name', 
    'html_url', 
    'description', 
    'url', 
    'labels_url', 
    'created_at', 
    'updated_at', 
    'pushed_at', 
    'size', 
    'stargazers_count', 
    'watchers_count', 
    'language', 
    'has_issues', 
    'has_projects', 
    'has_downloads', 
    'has_wiki', 
    'has_pages', 
    'has_discussions', 
    'forks_count', 
    'open_issues_count', 
    'license', 
    'allow_forking', 
    'topics', 
    'visibility', 
    'forks', 
    'open_issues', 
    'watchers', 
    'default_branch', 
    'score'
]

### Create two paths to save .json

# file with repos up to 21947 stars
file_no_range = '../data/df_repos_metadata_star_up_to_max.json'
# file with repos 0 to 21947 stars
file_range = '../data/df_repos_metadata_star_0_to_21947.json'

### Get column names for metadata file

help_columns = load_json(path='../data/helper/help_columns.json')
keys = list(help_columns[0].keys()) # get keys of loaded_data as list

len(keys)

### Get paths of raw data

# loaded zip files of metadata in raw_data
path_no_range = '../data/raw_data_zip/raw_data_no_range.zip'
path_range = '../data/raw_data_zip/raw_data_range_0_22196.zip' # evtl. in 0_212947 umwandeln. ich glaube, ich habe mich beim benennen mal vertan und einen alten filenamen nicht akutalisiert

# was executed before
# rezip_files(path=path_no_range, range_type='no_range')
# rezip_files(path=path_range, range_type='range')

path_range = Path('../data/raw_data/range') 
all_files_range = [file.name for file in path_range.iterdir() if file.is_file()]

path_no_range = Path('../data/raw_data/no_range') 
all_files_no_range = [file.name for file in path_no_range.iterdir() if file.is_file()]

### Open all jsons and load repo metadata into dataframe

# create empty df with keys of loaded_data as columns
df_raw_range = pd.DataFrame(columns=keys)
df_raw_no_range = pd.DataFrame(columns=keys)

for file in all_files_range:
    data = load_json(path=f'../data/raw_data/range/{file}')
    #  iterate through subdictionary in data and concatenate the content of the subdictionary to df_repos
    for repo in data:
        # create tmp df_repo for each repo
        df_tmp = pd.DataFrame(data=[repo], columns=keys)
        # concatenate df_repos with df_repo
        df_raw_range = pd.concat([df_raw_range, df_tmp], ignore_index=True)

for file in all_files_no_range:
    data = load_json(path=f'../data/raw_data/no_range/{file}')
    data = data['items']
    #  iterate through subdictionary in data and concatenate the content of the subdictionary to df_repos
    for repo in data:
        # create tmp df_repo for each repo
        df_tmp = pd.DataFrame(data=[repo], columns=keys)
        # concatenate df_repos with df_repo
        df_raw_no_range = pd.concat([df_raw_no_range, df_tmp], ignore_index=True)

### Clean remove unnecessary columns of df_raw_range and df_raw_no_range

for k in keys: # iterate over all keys
    if k in columns: # check if key is in list with columns
        continue
    else:
        df_raw_range = df_raw_range.drop([k], axis=1) # if not drop it

df_raw_range.info()

for k in keys: # iterate over all keys
    if k in columns: # check if key is in list with columns
        continue
    else:
        df_raw_no_range = df_raw_no_range.drop([k], axis=1) # if not drop it

df_raw_no_range['stargazers_count'].describe() # check info of df

### Save each dataframe in a single JSON file

tmp_json = df_raw_range.to_json(orient='records', lines=False, force_ascii=False)

# '../data/df_repos_metadata_star_0_to_22196.json'
with open(file_range, 'w') as file:
   file.write(tmp_json)

tmp_json = df_raw_no_range.to_json(orient='records', lines=False, force_ascii=False)

# '../data/df_repos_metadata_star_up_to_max.json'
with open(file_no_range, 'w') as file:
   file.write(tmp_json)

### Load each JSON file and prepare it for concatinatig

with open('../data/df_repos_metadata_star_up_to_max.json', 'r') as file: # metadata from first request with no constraints for star count
    loaded_data = json.load(file)

df1 = pd.DataFrame(loaded_data)

df1.info()

df1.describe() # compute metrics of df1

quantile_75 = df1['stargazers_count'].describe().loc['75%'] # save 75th percentile
quantile_75

len(df1[df1['stargazers_count'] > quantile_75]) # check length

df1 = df1[df1['stargazers_count'] > quantile_75] # filter df1 so it only contains repos with star count > 75th percentile 

len(df1) # check length

with open('../data/df_repos_metadata_star_0_to_21947.json', 'r') as file: # metadata from second request with for star count
    loaded_data = json.load(file)

df2 = pd.DataFrame(loaded_data)

df2.info() # get info about df2

df2.describe() # compute metrics

### Combine the two dataframes into one for further analysis

df_repos = pd.DataFrame() # create new empty df_repos
df_repos = pd.concat([df2, df1]) # fill df_repos with df1 and df2

df_repos.info() # get info about df_repos

### Remove duplicates

# the behavior of the api is sometimes confusing, there are some duplicates in the dataframe --> remove them
df_cleaned = df_repos.drop_duplicates(subset=['full_name'])

### Filter df_repos for spoken language 'English'

def detect_language(text):
    try:
        return detect(text)
    except:
        return "error"

df_repos['language_spoken'] = df_repos['description'].apply(detect_language) # create new col with language and add data

df_repos['language_spoken'].value_counts() # check language distribution

len(df_repos[df_repos['language_spoken'] == 'en']) # check number of rows where description is written in English

### Save final file for the repository metadata as JSON for further processing

tmp_json = df_repos.to_json(orient='records', lines=False, force_ascii=False)

with open('../data/df_repos_metadata.json', 'w') as file:
   file.write(tmp_json)

### Load JSON with repository metadata (test)

with open('../data/df_repos_metadata.json', 'r') as file:
    loaded_data = json.load(file)

df = pd.DataFrame(loaded_data)

df.info()
