==> preprocessing/03_get_split_of_repos.ipynb <==
**Description:** In this notebook a file 'df_repos_metadata.json' will be loaded and it's content processed to create a sample of GitHub repository metadata for further processing. For the sample creation the correct spoken language and license type are considered. The created dataframe (df_repos_sample_250.json) is used for the collection of the entire content of GitHub repository (see file /github_api/02_get_repo_data.py).

*Hint: If lines are created with support of a Large Language Model or the code is taken from another source, you find following hint at the end of the line: (generated with Microsoft Copilot) or (source: link_to_source)*

import pandas as pd # package for data manipulation
import json # package to work with .json

### Load data

file_path = '../data/df_repos_metadata.json' # filepath of input file
with open(file_path, 'r') as file: # JSON file is loaded and it contents saved within a variable
    loaded_data = json.load(file)

df = pd.DataFrame(data=loaded_data) # create dataframe from previous loaded JSON

len(df) # check length of dataframe

### Filter dataframe with specific conditons which are requiered for further processing

# filter repos for spoken language English and save the data in a new dataframe
# saving the language filter was forgotten in notebook 01_repos_metadata.ipynb!!!
df = df[df['language_spoken'] == 'en']

len(df) # check length of dataframe

df['language_spoken'].value_counts() # check if dataframe only contains repositories with the value 'en' for the column 'language_spoken'

# add new column with license type of each repository
df['license_type'] = df['license'].apply(lambda x: list(x.values())[0] if isinstance(x, dict) else None) # (generated with Microsoft Copilot)

df.head(2)

# list with licenses which are allowing to use the content of the repository for further processing
licenses_list = ['apache-2.0', 'mit', 'isc', 'bsd-3-clause', 'bsd-2-clause'] 

df = df[df['license_type'].apply(lambda x: x in licenses_list)] # filter df for licenses_types which are in licenses_list and save dataframe

len(df) # check length of dataframe

df['license_type'].value_counts() # check which licenses_type the dataframe contains

### Create sample

sample_df = df.sample(n=250) # create a sample of 250 elements

len(sample_df) # check length of dataframe

### Save sample in JSON file

tmp_json = sample_df.to_json(orient='records', lines=False, force_ascii=False) # save the sample dataframe in variable tmp_json
with open('../data/df_repos_sample_250.json', 'w') as file: # write content fo tmp_json to new JSON file
    file.write(tmp_json)
